Refactor UriIndicator
